docs: add LAYOUT.md for player app shared layout composables
Documents SpScreen, SpScrollableContent, SpMainContentPadding,
SpScreenTopSpacer, SpSectionList, and SpCarousel with usage patterns
and rules. Referenced from CLAUDE.md.

Also: migrate HomeScreen and LibraryConsolesTab to use SpScreen
with gradient support, add gradient background parameter to SpScreen.
