+++ player/shared/src/commonMain/kotlin/com/spela/player/presentation/ui/feature/library/LibraryConsolesTab.kt
@@ -1,13 +1,9 @@
 package com.spela.player.presentation.ui.feature.library
 
-import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.BoxWithConstraints
-import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.lazy.rememberLazyListState
 import androidx.compose.ui.platform.testTag
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.pulltorefresh.PullToRefreshBox
@@ -17,17 +13,14 @@ import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.drawBehind
-import androidx.compose.ui.geometry.Offset
-import androidx.compose.ui.graphics.Brush
 import androidx.compose.ui.unit.dp
 import com.spela.player.presentation.intent.GameListIntent
 import com.spela.player.presentation.ui.components.SpEmptyStates
-import com.spela.player.presentation.ui.gamepad.InputMode
-import com.spela.player.presentation.ui.gamepad.LocalInputMode
-import com.spela.player.presentation.ui.theme.LocalTitleBarInset
+import com.spela.player.presentation.ui.components.SpMainContentPadding
+import com.spela.player.presentation.ui.components.SpScreen
+import com.spela.player.presentation.ui.components.SpScreenTopSpacer
+import com.spela.player.presentation.ui.components.SpScrollableContent
 import com.spela.player.presentation.ui.theme.SpColor
-import com.spela.player.presentation.ui.theme.SpSpacing
 import com.spela.player.presentation.viewmodel.GameListViewModel
 
 @OptIn(ExperimentalMaterial3Api::class)
@@ -42,41 +35,18 @@ internal fun LibraryConsolesTab(
         viewModel.onIntent(GameListIntent.LoadConsoles)
     }
 
-    val titleBarInset = LocalTitleBarInset.current
     val gradientColors = listOf(
         SpColor.PrimaryDark.darken(0.75f),
         SpColor.AccentDark.darken(0.80f),
     )
 
-    Box(
-        modifier = Modifier
-            .fillMaxSize()
-            .drawBehind {
-                val cx = size.width / 2f
-                val cy = size.height / 2f
-                val d = (size.width + size.height) * 0.25f
-                drawRect(
-                    brush = Brush.linearGradient(
-                        colors = gradientColors,
-                        start = Offset(cx - d, cy - d),
-                        end = Offset(cx + d, cy + d),
-                    ),
-                )
-            },
-    ) {
+    SpScreen(gradientColors = gradientColors) {
         if (state.isLoading && state.consoles.isEmpty()) {
-            BoxWithConstraints(modifier = Modifier.fillMaxSize()) {
-                val columnsPerRow = if (maxWidth >= 600.dp) 2 else 1
-                LazyColumn(
-                    contentPadding = PaddingValues(
-                        start = SpSpacing.ScreenHorizontal,
-                        end = SpSpacing.ScreenHorizontal,
-                        top = titleBarInset + SpSpacing.Default,
-                        bottom = SpSpacing.Default,
-                    ),
-                    verticalArrangement = Arrangement.spacedBy(SpSpacing.Medium),
-                ) {
-                    item {
+            SpScrollableContent {
+                SpScreenTopSpacer()
+                SpMainContentPadding {
+                    BoxWithConstraints(modifier = Modifier.fillMaxWidth()) {
+                        val columnsPerRow = if (maxWidth >= 600.dp) 2 else 1
                         ConsolesSkeletonGrid(
                             columnsPerRow = columnsPerRow,
                             count = if (columnsPerRow == 2) 4 else 3,
@@ -98,23 +68,9 @@ internal fun LibraryConsolesTab(
                         SpEmptyStates.EmptyLibrary()
                     }
                 } else {
-                    val listState = rememberLazyListState()
-                    LazyColumn(
-                        state = listState,
-                        modifier = Modifier.fillMaxSize().testTag("consoles-list"),
-                        contentPadding = run {
-                            val isGamepad = LocalInputMode.current == InputMode.GAMEPAD
-                            val topPad = if (isGamepad) SpSpacing.SectionIndicatorClearance else titleBarInset + SpSpacing.Default
-                            PaddingValues(
-                                start = SpSpacing.ScreenHorizontal,
-                                end = SpSpacing.ScreenHorizontal,
-                                top = topPad,
-                                bottom = SpSpacing.Default,
-                            )
-                        },
-                        verticalArrangement = Arrangement.spacedBy(SpSpacing.Medium),
-                    ) {
-                        item {
+                    SpScrollableContent(modifier = Modifier.testTag("consoles-list")) {
+                        SpScreenTopSpacer()
+                        SpMainContentPadding {
                             BoxWithConstraints(modifier = Modifier.fillMaxWidth()) {
                                 val columnsPerRow = when {
                                     maxWidth >= 700.dp -> 3
